requirements.txt
paho-mqtt
pyserial
scikit-fuzzy
numpy
networkx
logging
datetime
time